# src/main/resources/swagger/api.yml
...
openapi: 3.0.1
info:
  title: thg-procure-questionnaire-service
  description: This is to serve as the documentation of the questionnaire-service APIs for THG Procure.
  version: 1.0.0
  contact:
    name: Abhijeet Singh
    email: abhijeet.singh@cloudesign.com
servers:
  - url: https://dev.qts.thg-procure.com
  - url: https://uat.qts.thg-procure.com
  - url: https://qts.thg-procure.com
paths:
...
  /client:
    description: API endpoints to manage Clients of Questionnaire Service
    post:
      description: Create a new Client
      parameters:
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createClient'
      responses:
        200:
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClient'
        400:
          description: Bad Request
        401:
          description: Unauthorized
  /client/{id}:
    put:
      description: Update an existing Client
      parameters:
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: path
          name: id
          description: Uniquely identifies the client.
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateClient'
      responses:
        200:
          description: Success Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getClient'
        400:
          description: Bad Request
        404:
          description: Requested Client id not found
        401:
          description: Unauthorized
    delete:
      description: Disable a Client
      parameters:
        - in: header
          name: qts-request-id
          description: Uniquely identifies the request.
          required: false
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
        - in: path
          name: id
          description: Uniquely identifies the client to the Questionnaire Service
          required: true
          schema:
            type: string
            format: uuid
            example: 7ffc1b3a-5395-11ee-8c99-0242ac120002
      responses:
        204:
          description: Deleted
        404:
          description: Requested Client id not found
        401:
          description: Unauthorized
        400:
          description: Bad Request
components:
  schemas:
...
    createClient:
      type: object
      required: [clientName, description]
      properties:
        clientName:
          type: string
        description:
          type: string
    updateClient:
      type: object
      required: [clientName, description]
      properties:
        clientName:
          type: string
        description:
          type: string
    getClient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        clientName:
          type: string
        description:
          type: string